[FIX] added hooks,modified version,MVC structure and simple headers
