--- partner_address_street3/model/partner.py
@@ -1,48 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Author: Nicolas Bessi
-#    Copyright 2014-2015 Camptocamp SA
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-from openerp import models, fields, api
-
-
-class ResPartner(models.Model):
-    """Add third field in address"""
-
-    _inherit = "res.partner"
-    street3 = fields.Char('Street 3')
-
-    @api.model
-    def _address_fields(self):
-        fields = super(ResPartner, self)._address_fields()
-        fields.append('street3')
-        return fields
-
-
-class res_country(models.Model):
-    """Override default adresses formatting of countries"""
-
-    _inherit = 'res.country'
-
-    address_format = fields.Text(
-        default=(
-            "%(street)s\n%(street2)s\n%(street3)s\n"
-            "%(city)s %(state_code)s %(zip)s\n"
-            "%(country_name)s"
-        )
-    )
