+++ partner_address_street3/models/__init__.py
@@ -3,6 +3,5 @@
 # Copyright 2016 Sodexis
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import models
-from .hooks import post_init_hook
-from .hooks import uninstall_hook
+from . import partner
+from . import res_country
