+++ partner_address_street3/models/partner.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+# Copyright 2014 Nicolas Bessi, Alexandre Fayolle, Camptocamp SA
+# Copyright 2016 Sodexis
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from openerp import models, fields, api
+
+
+class ResPartner(models.Model):
+    """Add third field in address"""
+
+    _inherit = "res.partner"
+    street3 = fields.Char('Street 3')
+
+    @api.model
+    def _address_fields(self):
+        fields = super(ResPartner, self)._address_fields()
+        fields.append('street3')
+        return fields
