+++ partner_address_street3/view/partner_view.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-  <data>
+<odoo>
     <record id="add_stree3_in_form_view" model="ir.ui.view">
       <field name="name">add stree3 in form view</field>
       <field name="model">res.partner</field>
@@ -57,5 +56,4 @@
       </field>
     </record>
 
-  </data>
-</openerp>
+</odoo>
